requirements.txt
ipykernel
pandas
openpyxl
matplotlib
seaborn
plotly
tabulate
python-dotenv
langchain
langchain-google-genai
langchain_core
langchain_community
langgraph
streamlit
llama-index
llama-index-core
llama-index-llms-gemini